outsourced password reset to component
